+ Implemented design made by Thanus

+++ ShufflerTumblr/ShufflerTumblr/Classes/DashBoardViewController.m
@@ -157,7 +157,7 @@
 
 #pragma UITableView delegates
 -(void) tableView:(UITableView *) tableView didSelectRowAtIndexPath:(NSIndexPath *)indexPath {
-    chosenRow = indexPath.row;
+    chosenRow = (int)indexPath.row;
     [tableView deselectRowAtIndexPath:indexPath animated:YES];
